prototype/pob_server/requirements.txt
fastapi
uvicorn
websockets
google-genai
Pillow
numpy
requests
python-multipart
